pyproject.toml
[project]
name = "chat_uv"
version = "0.1.0"
description = "Telegram bot backend using FastAPI"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "python-telegram-bot",
  "python-dotenv",
]